+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 - Arguments:
   - '--confirma-category'
+- Warning when trying to run tests without '--confirma-run' argument.
 
 ## [0.8.1-beta 2024-08-20]